Fixed class inheritance
