+++ gptables/__init__.py
@@ -1,7 +1,8 @@
 from gptables.core.theme import Theme
 from gptables.core.gptable import GPTable
-from gptables.core.workbook_wrapper import Workbook
-from gptables.core.worksheet_wrapper import Worksheet
+from gptables.core.wrappers import GPWorksheet
+from gptables.core.wrappers import GPWorkbook
+
 
 from gptables.core.api import (
         #datatypes
